+++ doc/CHANGES.rst
@@ -35,7 +35,8 @@ Restructuring
 - Removed experimental support for configuring the Twisted HTTP server
   as an alternative to ``ZServer``.
 
-- Moved ``Products/Five/security.py`` into the AccessControl package.
+- Moved ``Products/Five/security.py`` and security related ZCML configuration
+  into the AccessControl package.
 
 - Moved ``Products/Five/traversing.zcml`` directly into the configure.zcml.